+++ src/services/facades/Viewer.ts
@@ -1,0 +1,5 @@
+import { facade } from '@/framework/core/facades';
+
+import ViewerService from '@/services/ViewerService';
+
+export default facade(new ViewerService);